Loss Carryover - US
These changes reflect the NOL changes made in the CARES Act. I based the changes on a piece by Garrett (https://taxfoundation.org/phase-4-relief-legislation-net-operating-loss-cares-act/). Since the new 5-year carryback also applies to tax years 2018 and 2019, I made the changes for the three years 2018 to 2020. Also, I deleted the 80% limit on carryforwards for 2020 (as far as I understand it only applies to 2020 and not to 2018 and 2019).