Update Variable vat_base
Calculated the 2017 and 2018 VAT Revenue Ratio (VRR) for all OECD countries and added a 2-year lag. Previously, the 2016 VRR was used for the years 2017, 2018, and 2019. No VAT revenue data was available for Australia and Greece for the year 2018, so the 2017 data was copied for those two countries. Details regarding the calculation of the VRR can be found in the Source Documentation.