Japan Update 2014-2019
Enterprise tax and business premises tax are deductible in the calculation of the taxable income for corporation tax purposes on a cash basis. However, corporation tax and inhabitant’s tax are not deductible. Fixed assets tax and other taxes are deductible, when assessed. Foreign income taxes also may be deductible if the Japanese corporation does not elect to claim a foreign tax credit.