Rename Variables in "Paying Taxes"
R was automatically renaming column headings in the "paying taxes" data because they were not unique for each column. However, the renaming was done in a different way on Daniel's vs. Elke's computer, so the code kept breaking. To fix the problem, the column headings were changed directly in the xlsx file.