Update Paying Taxes Data
PwC/World Bank did not update their "paying taxes" report this year due to the COVID-19 pandemic. The code was adjusted to now include the year 2011 as well and then to adjust the time frame to have data for each year. Because of that there is now a 3-year lag (previously a 2-year lag) in the data.