Update Capital Allowances Data
Use the updated code and dataset from the TF publication "Capital Cost Recovery in the OECD" to update the data here. The data for 2020 was added for all countries (there is a 1-year lag), Colombia's data was added for all years, and there were a few minor corrections (UK, NZ).