Update 2021 VAT Base
Did the 2021 VRR calculations for all OECD countries and the VRR calculations for all years for Colombia. Downloaded final consumption data and VAT/sales tax revenue from OECD stats and updated the calculations for all years (the OECD made small changes to many countries' VAT/sales tax revenue and final consumption data all the way back to 2014). So the VRRs shown in previous editions of the Index may be slightly different than the ones in the 2021 edition. The OECD's 2020 Consumption tax report has updated VRR calculations, but they end in 2018. Calculating the VRRs ourselves allows us to have them up to 2019, so there is only a 2-year (instead of a 3-year) lag. Also, updated the related source documentation.