Correction: Participation Exemption Israel and Korea
The participation exemption was incorrect for Israel and Korea. Since Colombia's participation exemption for holding companies is shown in the Index, Israel's is now also shown (unlike in previous years). There were no changes in Korea's participation exemption (the one they have is only for domestic companies)--this was corrected.