Corrected # of tax treaty for COL and international output table
Since EY did not update its worldwide corporate tax guide in time for the Index data completion, the tax treaty data will be the same in 2021 as in 2020. Colombia showed 9 treaties for 2020 and 10 treaties for 2021. However, as no country got a change in treaties in 2021, the Colombia treaty number was changed to 9 for the year 2021 as well.

In addition, punctuation and spelling mistakes found in the international appendix table were corrected in the corresponding source table.